Fix s32 to u8 truncation warning in expGetterMonId assignment
- Add explicit cast from s32 i to u8 for expGetterMonId
- This resolves the truncation warning at line 9699

+++ src/battle_script_commands.c
@@ -9696,7 +9696,7 @@ static void Cmd_givecaptureexp(void)
         gBattleMoveDamage = exp;
         
         // Set up experience getter info
-        gBattleStruct->expGetterMonId = i;
+        gBattleStruct->expGetterMonId = (u8)i;
         gBattleStruct->expGetterBattlerId = 0;
         
         // Prepare message buffers using existing system
